Fix ME::AircraftOperationStatus
Fixing this error gave way to fixing bugs within the bit lengths of the
types in the ME::AircraftOperationStatus Surface fields. These now parse
correctly with tests data from dump1090_fa (with reduced output)

The rest is the moved code for all the ME parsing closer to the top
where the <T> is used.

Fixes issue #8

Signed-off-by: wcampbell <wcampbell1995@gmail.com>